+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added buttons linking between Notifications and log.
 - Added action to "Save and add another".
 - Added action to "Delete notification".
+- Added trigger event ["When an entry is saved (per each site)"](https://plugins.doublesecretagency.com/notifier/events/types/entries).
 
 ### Changed
 - New filtering mechanism borrowed heavily from the [Webhooks](https://github.com/craftcms/webhooks/blob/9c901be4d98c8584893c1cbce2b7dad217fbc480/README.md#filtering-events) plugin.